refactor(publish): reorganize `publish` CLI subcommand
