+++ rust/common/Cargo.toml
@@ -10,7 +10,7 @@ async-trait = "0.1.84"
 bs58 = "0.5.1"
 chrono = { version = "0.4.39", features = ["serde"] }
 chrono-humanize = "0.2.3"
-clap = { version = "4.5.23", features = ["derive", "color"] }
+clap = { version = "4.5.23", features = ["derive", "color", "env", "wrap_help"] }
 derivative = "2.2.0"
 derive_more = "0.99.18"
 dirs = "5.0.1"
