+++ rust/publish-ghost/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "publish-ghost"
+version = "0.0.0"
+edition = "2021"
+
+[dependencies]
+common = { "path" = "../common" }
