+++ rust/publish-stencila/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "publish-stencila"
+version = "0.0.0"
+edition = "2021"
+
+[dependencies]
+cloud = { path = "../cloud" }
+codec = { "path" = "../codec" }
+codec-swb = { "path" = "../codec-swb" }
+common = { "path" = "../common" }
+document = { path = "../document" }
+schema = { "path" = "../schema" }
